added middleware traffic router
1) middleware traffic router in front of routes
2) changed capture route to properly store original urls, headers, and query params from req object
3) changed GET bin route to read back
4) updated schema & seeding files for new format
5) some clean up & organization
